backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pymupdf
pydantic
python-dotenv
python-multipart
google-genai
